backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0